+++ zcash_primitives/src/sapling/keys.rs
@@ -13,7 +13,7 @@ use super::{
     },
     spec::{
         crh_ivk, diversify_hash, ka_sapling_agree, ka_sapling_agree_prepared,
-        ka_sapling_derive_public,
+        ka_sapling_derive_public, ka_sapling_derive_public_subgroup_prepared, PreparedBaseSubgroup,
     },
 };
 use crate::{
@@ -24,7 +24,7 @@ use crate::{
 use blake2b_simd::{Hash as Blake2bHash, Params as Blake2bParams};
 use ff::PrimeField;
 use group::{Curve, Group, GroupEncoding};
-use subtle::{ConstantTimeEq, CtOption};
+use subtle::{Choice, ConditionallySelectable, ConstantTimeEq, CtOption};
 use zcash_note_encryption::EphemeralKeyBytes;
 
 /// Errors that can occur in the decoding of Sapling spending keys.
@@ -241,11 +241,9 @@ pub struct SaplingIvk(pub jubjub::Fr);
 
 impl SaplingIvk {
     pub fn to_payment_address(&self, diversifier: Diversifier) -> Option<PaymentAddress> {
-        diversifier.g_d().and_then(|g_d| {
-            let pk_d = g_d * self.0;
-
-            PaymentAddress::from_parts(diversifier, pk_d)
-        })
+        let prepared_ivk = PreparedIncomingViewingKey::new(self);
+        DiversifiedTransmissionKey::derive(&prepared_ivk, &diversifier)
+            .and_then(|pk_d| PaymentAddress::from_parts(diversifier, pk_d))
     }
 
     pub fn to_repr(&self) -> [u8; 32] {
@@ -259,6 +257,62 @@ pub struct Diversifier(pub [u8; 11]);
 impl Diversifier {
     pub fn g_d(&self) -> Option<jubjub::SubgroupPoint> {
         diversify_hash(&self.0)
+    }
+}
+
+/// The diversified transmission key for a given payment address.
+///
+/// Defined in [Zcash Protocol Spec § 4.2.2: Sapling Key Components][saplingkeycomponents].
+///
+/// Note that this type is allowed to be the identity in the protocol, but we reject this
+/// in [`PaymentAddress::from_parts`].
+///
+/// [saplingkeycomponents]: https://zips.z.cash/protocol/protocol.pdf#saplingkeycomponents
+#[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
+pub struct DiversifiedTransmissionKey(jubjub::SubgroupPoint);
+
+impl DiversifiedTransmissionKey {
+    /// Defined in [Zcash Protocol Spec § 4.2.2: Sapling Key Components][saplingkeycomponents].
+    ///
+    /// Returns `None` if `d` is an invalid diversifier.
+    ///
+    /// [saplingkeycomponents]: https://zips.z.cash/protocol/protocol.pdf#saplingkeycomponents
+    pub(crate) fn derive(ivk: &PreparedIncomingViewingKey, d: &Diversifier) -> Option<Self> {
+        d.g_d()
+            .map(PreparedBaseSubgroup::new)
+            .map(|g_d| ka_sapling_derive_public_subgroup_prepared(ivk.inner(), &g_d))
+            .map(DiversifiedTransmissionKey)
+    }
+
+    /// $abst_J(bytes)$
+    pub(crate) fn from_bytes(bytes: &[u8; 32]) -> CtOption<Self> {
+        jubjub::SubgroupPoint::from_bytes(bytes).map(DiversifiedTransmissionKey)
+    }
+
+    /// $repr_J(self)$
+    pub(crate) fn to_bytes(self) -> [u8; 32] {
+        self.0.to_bytes()
+    }
+
+    /// Returns true if this is the identity.
+    pub(crate) fn is_identity(&self) -> bool {
+        self.0.is_identity().into()
+    }
+
+    /// Exposes the inner Jubjub point.
+    ///
+    /// This API is exposed for `zcash_proof` usage, and will be removed when this type is
+    /// refactored into the `sapling-crypto` crate.
+    pub fn inner(&self) -> jubjub::SubgroupPoint {
+        self.0
+    }
+}
+
+impl ConditionallySelectable for DiversifiedTransmissionKey {
+    fn conditional_select(a: &Self, b: &Self, choice: Choice) -> Self {
+        DiversifiedTransmissionKey(jubjub::SubgroupPoint::conditional_select(
+            &a.0, &b.0, choice,
+        ))
     }
 }
 
@@ -290,8 +344,8 @@ impl EphemeralSecretKey {
         EphemeralPublicKey(ka_sapling_derive_public(&self.0, &g_d))
     }
 
-    pub(crate) fn agree(&self, pk_d: &jubjub::ExtendedPoint) -> SharedSecret {
-        SharedSecret(ka_sapling_agree(&self.0, pk_d))
+    pub(crate) fn agree(&self, pk_d: &DiversifiedTransmissionKey) -> SharedSecret {
+        SharedSecret(ka_sapling_agree(&self.0, &pk_d.0.into()))
     }
 }